# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.ArticleDetail:
    properties:
      author:
        $ref: '#/definitions/entity.ProfilePreview'
      body:
        type: string
      createdAt:
        type: string
      description:
        type: string
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      slug:
        type: string
      tagList:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  entity.CommentDetail:
    properties:
      author:
        $ref: '#/definitions/entity.ProfilePreview'
      body:
        type: string
      createdAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
    type: object
  entity.ProfilePreview:
    properties:
      bio:
        type: string
      followersCount:
        type: integer
      following:
        type: boolean
      image:
        type: string
      username:
        type: string
    type: object
...
  request.CommentCreate:
    properties:
      body:
        example: this is a comment
        maxLength: 10000
        type: string
    required:
    - body
    type: object
  request.CommentCreateRequest:
    properties:
      comment:
        $ref: '#/definitions/request.CommentCreate'
    type: object
...
  response.ArticleDetailResponse:
    properties:
      article:
        $ref: '#/definitions/entity.ArticleDetail'
    type: object
...
  response.CommentDetailResponse:
    properties:
      comment:
        $ref: '#/definitions/entity.CommentDetail'
    type: object
  response.CommentDetailsResponse:
    properties:
      articlesCount:
        type: integer
      comments:
        items:
          $ref: '#/definitions/entity.CommentDetail'
        type: array
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 18
        type: integer
    type: object
  response.Error:
    properties:
      error:
        example: message
        type: string
    type: object
  response.ProfilePreviewResponse:
    properties:
      profile:
        $ref: '#/definitions/entity.ProfilePreview'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Realworld API using Golang + Gin + Clean Architecture
  title: Realworld Fiber Clean API
  version: "1.0"
paths:
...
  /articles/{slug}/comments:
    get:
      description: Get all comments of an article
      operationId: comments-get-all
      parameters:
      - description: Limit number of results
        format: int64
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        format: int64
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommentDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create Comment
      operationId: comments-create
      parameters:
      - description: Create Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CommentDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Create Comment
      tags:
      - comments
  /articles/{slug}/comments/{commentID}:
    delete:
      description: Delete comment by id
      operationId: comment-delete-by-id
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - comments
  /articles/{slug}/favorite:
    delete:
      description: Unfavorite an article
      operationId: favorites-delete
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticleDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Unfavorite Article
      tags:
      - favorites
    post:
      consumes:
      - application/json
      description: Favorite an article
      operationId: favorites-create
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ArticleDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Favorite Article
      tags:
      - favorites
...
  /profiles/{username}:
    get:
      description: Get profile by username
      operationId: profiles-get-by-username
      parameters:
      - description: Username of the profile to get
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProfilePreviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - profiles
  /profiles/{username}/follow:
    delete:
      description: Unfollow user by username
      operationId: profiles-unfollow
      parameters:
      - description: Username of the profile to unfollow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProfilePreviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Unfollow user
      tags:
      - profiles
    post:
      description: Follow user by username
      operationId: profiles-follow
      parameters:
      - description: Username of the profile to follow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProfilePreviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Follow user
      tags:
      - profiles
...
securityDefinitions:
  BearerAuth:
    description: Type "Token" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"